refining flatlist
