+++ client/App/components/index.tsx
@@ -7,7 +7,8 @@ import City from './City';
 import Filters from './Filters';
 import ProfileItem from './ProfileItem';
 import ProfileModal from './ProfileModal';
-import Message from './MostRecentMessage';
+import RecentMessage from './RecentMessage';
+import AllChats from './AllChats';
 
 export {
   Icon,
@@ -16,5 +17,6 @@ export {
   Filters,
   ProfileItem,
   ProfileModal,
-  Message
+  RecentMessage,
+  AllChats
 };
