+++ client/App/pages/Chat.tsx
@@ -1,16 +1,11 @@
 import React, {
-    useState,
-    useEffect,
     useRef,
-    FC
+    FC,
+    useState
 } from "react";
 import {
-    Text,
-    TextInput,
     View,
-    ImageBackground,
-    TouchableOpacity,
-    FlatList
+    ImageBackground
 } from 'react-native';
 // import io from 'socket.io-client';
 // import moment from 'moment';
@@ -18,73 +13,18 @@ import {
 import styles, { DARK_GRAY } from "../../assets/styles";
 import {
     Icon,
-    Message
+    AllChats
 } from "../components";
-import DEMO from "../../assets/data/demo";
-import {
-    useAuth,
-    useChatState
-} from "../context";
-import {
-    HomeTabNavigatorParamList,
-    RootNavigatorParamsList,
-    MessageStackParamList
-} from "../types";
-import { CompositeNavigationProp, RouteProp, useRoute } from "@react-navigation/native";
-import { NativeStackNavigationProp } from "@react-navigation/native-stack";
-
-import { getUserChats } from "../utils";
-import AsyncStorage from "@react-native-async-storage/async-storage";
+import { useAuth, useChatState} from "../context";
 import { ActivityIndicator } from "react-native-paper";
-// import ChatMessages from '../components/Chat/ChatMessages';
-export interface ChatProps {
-    navigation: CompositeNavigationProp<NativeStackNavigationProp<HomeTabNavigatorParamList, 'Chat'>,
-        NativeStackNavigationProp<RootNavigatorParamsList>>;
-    route: RouteProp<MessageStackParamList>;
-}
 
-const Chat: FC<ChatProps> = ({ navigation, route }) => {
+const Chat: FC<any> = ({ navigation, route }) => {
     //state initialized
-    const [loggedUser, setLoggedUser] = useState<any>();
-    const auth = useAuth();
-    const { token, userId } = auth.authData;
-    const {
-        user,
-        selectedChat,
-        setSelectedChat,
-        chats,
-        setChats
-    } = useChatState();
+    const [fetchAgain, setFetchAgain] = useState<any>();
     //reference set for socketRef using useRef hook
     const socketRef = useRef();
 
-    const fetchChats = async () => {
-        getUserChats(userId, token)
-            .then(res => {
-                // gets all user messages
-                setChats(res.data);
-            })
-            .catch(e => {
-                // throw error
-                console.error(e.message);
-            });
-    }
-
-    const handleLoggedUser = async () => {
-        AsyncStorage.getItem('@auth')
-            .then((res: any) => {
-                setLoggedUser(JSON.parse(res));
-            })
-            .catch((e: any) => {
-                console.error('Failed to get key', e.message);
-            });
-    }
-
-    useEffect(() => {
-        fetchChats();
-        handleLoggedUser();
-    }, []);
-
+    const { user } = useChatState();
 
     //return jsx to render UI
     return (
@@ -94,26 +34,7 @@ const Chat: FC<ChatProps> = ({ navigation, route }) => {
         >
             <View style={styles.containerMessages}>
                 {
-                    chats ? (
-                        <FlatList
-                            data={chats}
-                            keyExtractor={(item) =>  item._id }
-                            renderItem={({ item, index }) => (
-                                <TouchableOpacity
-                                    onPress={() => navigation.navigate('Message', { thread: item })}
-                                >
-                                    {console.log('flatlist', item._id, 'hello')}
-                                    <Message
-                                        image={item.image}
-                                        name={item.name}
-                                        lastMessage={item.message}
-                                    />
-                                </TouchableOpacity>
-                            )}
-                        />
-                    ) : (
-                        <ActivityIndicator style={styles.indicator} size="large" color="#F0ABC1" />
-                    )
+                    user ? <AllChats /> : <ActivityIndicator style={styles.indicator} size="large" color="#F0ABC1" />
                 }
             </View>
         </ImageBackground >
