+++ client/App/utils/index.tsx
@@ -167,7 +167,8 @@ const getUserChats = async (userId: string, token: string) => {
  */
 const createChat = async (userId: string, matchedUserId: string, token: string) => {
   let userIds = [];
-  userIds.push(userId, matchedUserId);
+  userIds.push(matchedUserId, userId);
+  console.log(userIds);
   
   return axios.post(`${API_URL}/chats`, { userIds, token });
 }
